Run the container as the specified user.
Support the RUNAS variable to run as the specified user.

The docker container is used only during development, ie. with the work
directory always bind mounted to a directory on the developer's box.
This means we want the container to always run as the current user and
RUNAS to be mandatory.

Fix #596.