Fix duplicate pages in Elasticsearch's index.
Pages are indexed twice because they're configured as a collection
(https://github.com/italia/developers.italia.it/issues/261#issuecomment-588283052)
and we are doing it as a workaround to keep things tidy and use a _pages directory,
which Jekyll does not have by default (https://github.com/jekyll/jekyll/issues/920).

BUT! Our content in _pages is already in two subdirectories ("it" and "en") AND
we already have two dirs named like that in the root of the repo, so moving
everything there and not having _pages is even tidier.

Fix #261.