feat: add pages to guide through the reuse declaration procedure
Use localStorage to pass the repo URL to the new page,
so we're not vulnerable to abuse (phishing, social engineering)
by direct manipulation of the query string.

The state doesn't persist after a refresh of the page
(ie, a refresh will show the whole page and the link to the repo will disappear),
but I think it's a good tradeoff.

Fix #799.

Co-authored-by: libremente <surf@libremente.eu>