revert: remove .vercel/ignore-gh-pages.sh
Revert "chore: tell Vercel not to build the gh-pages branch".

The script is not really needed in master because it'd need
to be on the "gh-pages" branch for it to work.

Vercel fails building "gh-pages" with no big consequences anyway,
so let's leave it at that until they provide a better way to
ignore branches in the UI.

Refs: af53ae66fd384854cda401a2f1b6180157f7ab9a.