fix: glitch in related software thumbnails
owl-carousel clones the items to give the impression of circular
scrolling, but it doesn't carry over the the fallback image mechanism
on cloned items.

The scrollability of that section is not really obvious and showing 4 items
should be enough, so let's remove it. It's also another step towards the
removal of jQuery.

Fix #1015.