feat: move GitHub exported data to a different repo
Move the github_importer.rb run out of the build stage, to another repo
(italia/developers.italia.it-data) which takes care of periodically
exporting the data to that very repo.

In doing so, we isolate the required token to that single use repo,
decrease the build time, make Vercel previews fully work,
become more resilient to temporary GitHup API outages and make things
a little bit simpler for contributors.

Fix #684.