feat: add PNRR search to the search engine
Add the capability of searching for PNNR related software.

We came up with the convention that Italian publiccode.yml files can put
"tags" in "description.it.features" to mark software that help Public administrations
with the PNRR (https://www.governo.it/sites/governo.it/files/PNRR.pdf).

We have three kinds of tags:

1. "PNRR" as the general tag
   It can be searched by adding "pnrr=1" to the query string
2. "PNRR/Beneficiari/$TARGET_NAME" the target the software is made for,
   fe. "Scuole" (Schools), "Comuni".
   It can be searched by adding "pnrr_target=$TARGET_NAME" to the query
   string (case sensitive)
3. "PNRR/Misura/$NUMBER" the PNRR measure, fe "1.4.1"
   It can be searched by adding "pnrr_measure=$NUMBER" to the query
   string